2007-05-04  Brian Nickel <brian.nickel@gmail.com>
	* src/TagLib/Riff/AviStream.cs:
	* src/TagLib/Riff/BitmapInfoHeader.cs:
	* src/TagLib/Riff/WaveFormatEx.cs:
	* src/TagLib/Id3v2/Frames/RelativeVolumeFrame.cs:
	* src/TagLib/Mpeg/File.cs:
	* src/TagLib/Mpeg/VideoHeader.cs:
	* src/TagLib/ByteVector.cs:
	* src/TagLib/WavPack/StreamHeader.cs:
	* src/TagLib/Ape/Footer.cs:
	* src/TagLib/Asf/PaddingObject.cs:
	* src/TagLib/Asf/File.cs:
	* src/TagLib/Asf/HeaderObject.cs:
	* src/TagLib/Asf/StreamPropertiesObject.cs:
	* src/TagLib/Asf/FilePropertiesObject.cs:
	* src/TagLib/Asf/ContentDescriptionObject.cs:
	* src/TagLib/Asf/ExtendedContentDescriptionObject.cs:
	* src/TagLib/Asf/Object.cs:
	* src/TagLib/Asf/Guid.cs:
	* src/TagLib/Asf/ContentDescriptor.cs:
	* src/TagLib/Mpeg4/Boxes/IsoMovieHeaderBox.cs:
	* src/TagLib/Mpeg4/Boxes/IsoSampleEntry.cs:
	* src/TagLib/Mpeg4/Boxes/IsoChunkLargeOffsetBox.cs:
	* src/TagLib/Mpeg4/Boxes/IsoAudioSampleEntry.cs:
	* src/TagLib/Mpeg4/Boxes/AppleElementaryStreamDescriptor.cs:
	* src/TagLib/Mpeg4/Boxes/IsoVisualSampleEntry.cs:
	* src/TagLib/Mpeg4/AppleTag.cs:
	* src/TagLib/Mpeg4/BoxHeader.cs:
	* src/TagLib/Mpeg4/Box.cs:
	* src/TagLib/Ogg/Codecs/Theora.cs:
	* src/TagLib/Ogg/PageHeader.cs: Goodbye, short and long, hello ushort
	and ulong. This should avoid those pesky OverflowExceptions.

svn path=/trunk/taglib-sharp/; revision=76662

+++ src/TagLib/Asf/ContentDescriptor.cs
@@ -43,7 +43,7 @@ namespace TagLib.Asf
       private string     name;
       private string     sValue;
       private ByteVector bvValue;
-      private long       lValue;
+      private ulong      lValue;
       
       
       //////////////////////////////////////////////////////////////////////////
@@ -70,14 +70,14 @@ namespace TagLib.Asf
          this.lValue = value;
       }
 
-      public ContentDescriptor (string name, long value) : this ()
+      public ContentDescriptor (string name, ulong value) : this ()
       {
          this.name   = name;
          this.type   = DataType.QWord;
          this.lValue = value;
       }
 
-      public ContentDescriptor (string name, short value) : this ()
+      public ContentDescriptor (string name, ushort value) : this ()
       {
          this.name   = name;
          this.type   = DataType.Word;
@@ -88,7 +88,7 @@ namespace TagLib.Asf
       {
          this.name   = name;
          this.type   = DataType.Bool;
-         this.lValue = value ? 1 : 0;
+         this.lValue = (ulong)(value ? 1 : 0);
       }
 
       public string Name {get {return name;}}
@@ -142,20 +142,20 @@ namespace TagLib.Asf
 
          ByteVector v = Object.RenderUnicode (name);
          
-         ByteVector data = Object.RenderWord ((short) v.Count);
+         ByteVector data = Object.RenderWord ((ushort) v.Count);
          data.Add (v);
-         data.Add (Object.RenderWord ((short) type));
+         data.Add (Object.RenderWord ((ushort) type));
          
          switch (type)
          {
             case DataType.Unicode:
                v = Object.RenderUnicode (sValue);
-               data.Add (Object.RenderWord ((short) v.Count));
+               data.Add (Object.RenderWord ((ushort) v.Count));
                data.Add (v);
             break;
             
             case DataType.Bytes:
-               data.Add (Object.RenderWord ((short) bvValue.Count));
+               data.Add (Object.RenderWord ((ushort) bvValue.Count));
                data.Add (bvValue);
             break;
             
@@ -176,7 +176,7 @@ namespace TagLib.Asf
             
             case DataType.Word:
                data.Add (Object.RenderWord (2));
-               data.Add (Object.RenderWord ((short) lValue));
+               data.Add (Object.RenderWord ((ushort) lValue));
             break;
             
             default:
