+++ src/TagLib/Asf/FilePropertiesObject.cs
@@ -29,12 +29,12 @@ namespace TagLib.Asf
       // private properties
       //////////////////////////////////////////////////////////////////////////
       private Guid file_id;
-      private long file_size;
-      private long creation_date;
-      private long data_packets_count;
-      private long play_duration;
-      private long send_duration;
-      private long preroll;
+      private ulong file_size;
+      private ulong creation_date;
+      private ulong data_packets_count;
+      private ulong play_duration;
+      private ulong send_duration;
+      private ulong preroll;
       private uint flags;
       private uint minimum_data_packet_size;
       private uint maximum_data_packet_size;
@@ -87,12 +87,12 @@ namespace TagLib.Asf
       // public properties
       //////////////////////////////////////////////////////////////////////////
       public Guid     FileId {get {return file_id;}}
-      public long     FileSize {get {return file_size;}}
-      public DateTime CreationDate {get {return new DateTime (creation_date);}}
-      public long     DataPacketsCount {get {return data_packets_count;}}
-      public TimeSpan PlayDuration {get {return new TimeSpan (play_duration);}}
-      public TimeSpan SendDuration {get {return new TimeSpan (send_duration);}}
-      public long     Preroll {get {return preroll;}}
+      public ulong    FileSize {get {return file_size;}}
+      public DateTime CreationDate {get {return new DateTime ((long)creation_date);}}
+      public ulong     DataPacketsCount {get {return data_packets_count;}}
+      public TimeSpan PlayDuration {get {return new TimeSpan ((long)play_duration);}}
+      public TimeSpan SendDuration {get {return new TimeSpan ((long)send_duration);}}
+      public ulong     Preroll {get {return preroll;}}
       public uint     Flags {get {return flags;}}
       public uint     MinimumDataPacketSize {get {return minimum_data_packet_size;}}
       public uint     MaximumDataPacketSize {get {return maximum_data_packet_size;}}
